server/requirements.txt
flask
flask-cors
pdfplumber
pytesseract
pdf2image
Pillow 